test(e2e): Refactor user authentication
